MIST-104: Change singular package names to plural
