+++ src/main/java/edu/snu/mist/task/DefaultPhysicalToChainedPlanImpl.java
@@ -15,7 +15,7 @@
  */
 package edu.snu.mist.task;
 
-import edu.snu.mist.task.operator.Operator;
+import edu.snu.mist.task.operators.Operator;
 
 import javax.inject.Inject;
