+++ src/main/java/edu/snu/mist/task/executor/queues/SchedulingQueue.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.mist.task.executor.queue;
+package edu.snu.mist.task.executor.queues;
 
 import org.apache.reef.tang.annotations.DefaultImplementation;
